requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==2.2.3
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.2
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.24.0
rq==1.7.0
selenium
six
soupsieve==2.0.1
urllib3==1.25.11
webdriver-manager==3.3.0
Werkzeug==1.0.1
wincertstore==0.2
cassandra-driver
plotly
